Rename data scrapers to match ./news
